[Linker] Fix Trampoline code breaking DWARF LPT
When trampolines are needed (e.g. Darwin ARM64), the DWARF LPT
generation fails because the replacement symbols are marked as external
symbols and skipped during the DWARF LPT generation phase.

This PR is still subject to extensive testing but has rectified the
issue for the very large binaries we link at Uber.